backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AquaGov FastAPI ML Microservice for Drought Prediction"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
    "tensorflow>=2.16.0",
    "h5py>=3.11.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.4.0",
    "pydantic>=2.7.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "twilio>=9.0.0",
    "apscheduler>=3.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]
